+++ driver/lib/sessions/android.ts
@@ -42,9 +42,10 @@ export async function startAndroidSession(
 export async function connectAndroidSession (
   this: FlutterDriver,
   androiddriver: AndroidUiautomator2Driver,
-  caps: Record<string, any>
+  caps: Record<string, any>,
+  clearLog: boolean = false
 ): Promise<IsolateSocket> {
-  const observatoryWsUri = await getObservatoryWsUri.bind(this)(androiddriver, caps);
+  const observatoryWsUri = await getObservatoryWsUri.bind(this)(androiddriver, caps, clearLog);
   return await connectSocket.bind(this)(observatoryWsUri, caps);
 }
 
@@ -52,7 +53,14 @@ export async function getObservatoryWsUri (
   this: FlutterDriver,
   proxydriver: AndroidUiautomator2Driver,
   caps: StringRecord,
+  clearLog: boolean = false
 ): Promise<string> {
+  if (clearLog) {
+    this._logmon?.clearlastMatch();
+    this._logmon?.stop();
+    this._logmon?.start();
+  }
+
   let urlObject: URL;
   if (caps.observatoryWsUri) {
     urlObject = new URL(caps.observatoryWsUri);
@@ -69,14 +77,18 @@ export async function getObservatoryWsUri (
         `Have you disabled it in capabilities?`
       );
     }
-    if (!this._logmon.lastMatch) {
+    const lastMatch = await this._logmon.waitForLastMatchExist(
+      caps.maxRetryCount,
+      caps.retryBackoffTime
+    );
+    if (!lastMatch) {
       throw new Error(
         `No observatory URL matching to '${OBSERVATORY_URL_PATTERN}' was found in the device log. ` +
         `Please make sure the application under test is configured properly according to ` +
         `https://github.com/appium/appium-flutter-driver#usage and that it does not crash on startup.`
       );
     }
-    urlObject = extractObservatoryUrl(this._logmon.lastMatch) as URL;
+    urlObject = extractObservatoryUrl(lastMatch) as URL;
   }
   const remotePort = urlObject.port;
   this.portForwardLocalPort = caps.forwardingPort ?? remotePort;