+++ driver/CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## 2.9.1
-- Fix observatory url connection after the app activation command
-- Bring `appium:maxRetryCount` and `appium:retryBackoffTime` back
+- Fix observatory url finding after an app activation
+    - Bring `appium:maxRetryCount` and `appium:retryBackoffTime` back to use for the observaotry url findings.
 
 ## 2.9.0
 - Tune syslog scanning to find the observatory url